message-brokers/event-driven-systems/order_api_kafka_pg/requirements.txt

Flask==3.0.3
kafka-python-ng==2.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
